+++ src/main/java/application/controller/HomeController.java
@@ -3,7 +3,7 @@ package application.controller;
 import webserver.annotation.MyController;
 import webserver.annotation.MyRequestMapping;
 import webserver.http.MyRequestMethod;
-import webserver.request.MyHttpRequest;
+import webserver.http.request.MyHttpRequest;
 
 @MyController
 public class HomeController {
