+++ src/main/java/webserver/http/request/MyHttpRequest.java
@@ -1,4 +1,4 @@
-package webserver.request;
+package webserver.http.request;
 
 import java.io.BufferedReader;
 import java.io.IOException;
